init: Lie about verified boot state being GREEN instead of YELLOW
Google's SafetyNet basic integrity check views this property among many others. On unmodified GrapheneOS, this check passes anyway despite the state being YELLOW, but the end goal is to also pass the CTS profile check while the bootloader is locked - so this makes sure nothing appears out of the ordinary.